changed camgains to be stored as float

+++ apps/sisoRec/config.h
@@ -16,7 +16,7 @@ namespace fs = boost::filesystem;
 struct CameraSettings
 {
 	int exposure;
-	int gain;
+	float gain;
 	bool displayed;
 };
